odds/requirements.txt
Flask==1.1.1
Flask-Caching==1.8.0
flask-cors==3.0.8
flask-restplus==0.13.0
flask-sqlalchemy-2.4.1
psycopg2-binary==2.8.4
SQLAlchemy==1.3.12
webargs==5.5.2
